requirements.txt
apache-airflow==3.0.1
beautifulsoup4==4.13.4
pandas==2.3.0
pendulum==3.1.0
selenium==4.33.0
